+++ captived/integration-tests/mode_Test.py
@@ -15,6 +15,7 @@ URL = 'http://[::1]:4000/mode'
 HEADERS = {'Content-Type':'application/json'}
 # use this directory as the root path for this test
 DATA_PATH = os.path.join(TESTDIR, 'config', 'default')
+PASSTHROUGH_PATH = os.path.join(TESTDIR, 'config', 'passthrough')
 
 ################################################################################
 ### mode_Test
@@ -78,6 +79,43 @@ class mode_Test(test.SharedServer, test.IntegrationTestCase):
         resp = requests.get(URL)
         self.assertMatchesFirstLineOfFile(router_mode, resp.json())
         
+
+
+################################################################################
+### mode_Test_passthrough
+### This test is nearly identical to the mode_Test, but it tests being able to
+### use an alternate symbolic link to the configuration directory.
+### Since the mode_Test already checked the functionality, all this does is
+### test reading the from the alternate directory.
+################################################################################
+class mode_Test_passthrough(test.SharedServer, test.IntegrationTestCase):
+
+    @classmethod
+    def setUpClass(cls, args=None):
+        super(mode_Test_passthrough, cls).setUpClass(['-p', PASSTHROUGH_PATH])
+
+
+    def setUp(self):
+        super(mode_Test_passthrough, self).setUp()
+        # Wait for the embedded server to start up.
+        time.sleep(1.1)
+
+    def tearDown(self):
+        # not changing in this test, so don't need to set mode back to default
+        super(mode_Test_passthrough, self).tearDown()
+
+    # This should get back the defult mode - which is passthrough for this test
+    def test_get_default_mode(self):
+        resp = requests.get(URL)
+        self.assertEqual(resp.json(), "passthrough")
+
+
+    def test_get_mode(self):
+        router_mode = os.path.join(PASSTHROUGH_PATH, 'data', 'systemd', 'xbridgex.target.active', 'id')
+        resp = requests.get(URL)
+        self.assertMatchesFirstLineOfFile(router_mode, resp.json())
+        
+
 
 if __name__ == '__main__':
     unittest.main(verbosity=2)
